Remove VirtualMESyncSlot
